INC-126 refactor: move pagination logic to hook, update Followers and Payments components
